Move instrument response out to separate class to avoid error.
Error was Unable to update extended plugin for element named with class not found.
