+++ base/uk.ac.stfc.isis.ibex.synoptic/META-INF/MANIFEST.MF
@@ -14,8 +14,8 @@ Require-Bundle: org.eclipse.core.runtime,
  uk.ac.stfc.isis.ibex.instrument,
  uk.ac.stfc.isis.ibex.json;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.configserver;bundle-version="1.0.0",
- uk.ac.stfc.isis.ibex.epics.switching,
- uk.ac.stfc.isis.ibex.devicescreens,
+ uk.ac.stfc.isis.ibex.epics.switching;visibility:=reexport,
+ uk.ac.stfc.isis.ibex.devicescreens;visibility:=reexport,
  uk.ac.stfc.isis.ibex.targets;bundle-version="1.0.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ActivationPolicy: lazy
