+++ base/uk.ac.stfc.isis.ibex.ui.alarm/META-INF/MANIFEST.MF
@@ -1,6 +1,6 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: Test
+Bundle-Name: Alarm
 Bundle-SymbolicName: uk.ac.stfc.isis.ibex.ui.alarm;singleton:=true
 Bundle-Version: 1.0.0.qualifier
 Bundle-Activator: uk.ac.stfc.isis.ibex.ui.alarm.Alarms
