+++ base/uk.ac.stfc.isis.ibex.ui.synoptic/META-INF/MANIFEST.MF
@@ -4,37 +4,18 @@ Bundle-Name: Synoptic
 Bundle-SymbolicName: uk.ac.stfc.isis.ibex.ui.synoptic;singleton:=true
 Bundle-Version: 1.0.0.qualifier
 Bundle-Activator: uk.ac.stfc.isis.ibex.ui.synoptic.Activator
-Require-Bundle: org.eclipse.ui,
- org.eclipse.core.runtime,
- uk.ac.stfc.isis.ibex.preferences,
- uk.ac.stfc.isis.ibex.ui.perspectives,
+Require-Bundle: uk.ac.stfc.isis.ibex.ui.perspectives,
  uk.ac.stfc.isis.ibex.ui;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.synoptic;bundle-version="1.0.0",
- uk.ac.stfc.isis.ibex.ui.widgets,
  uk.ac.stfc.isis.ibex.model,
- org.eclipse.core.databinding,
- org.eclipse.core.databinding.property;bundle-version="1.4.100",
- org.eclipse.core.databinding.beans;bundle-version="1.2.200",
- org.eclipse.jface.databinding,
- org.eclipse.swt,
- org.eclipse.ui.ide,
- uk.ac.stfc.isis.ibex.ui.dae;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.instrument;bundle-version="1.0.0",
- uk.ac.stfc.isis.ibex.configserver;bundle-version="1.0.0",
- uk.ac.stfc.isis.ibex.devicescreens,
- uk.ac.stfc.isis.ibex.ui.devicescreens;bundle-version="1.0.0"
+ uk.ac.stfc.isis.ibex.ui.devicescreens;bundle-version="1.0.0",
+ uk.ac.stfc.isis.ibex.targets;bundle-version="1.0.0",
+ uk.ac.stfc.isis.ibex.ui.targets;bundle-version="1.0.0",
+ uk.ac.stfc.isis.ibex.logger;bundle-version="1.0.0"
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Export-Package: uk.ac.stfc.isis.ibex.ui.synoptic,
  uk.ac.stfc.isis.ibex.ui.synoptic.component,
  uk.ac.stfc.isis.ibex.ui.synoptic.views,
  uk.ac.stfc.isis.ibex.ui.synoptic.widgets
-Import-Package: org.apache.logging.log4j,
- org.csstudio.opibuilder.runmode,
- org.csstudio.opibuilder.util,
- uk.ac.stfc.isis.ibex.epics.switching,
- uk.ac.stfc.isis.ibex.epics.writing,
- uk.ac.stfc.isis.ibex.logger,
- uk.ac.stfc.isis.ibex.opis,
- uk.ac.stfc.isis.ibex.targets,
- uk.ac.stfc.isis.ibex.ui.targets
